Algorithmic Geometric Mean
I have the bones of the algorithm together, I have not tested it so far, but I am curious to see its performance. As of now to perform an iteration, you must manually call the cast from the manager. I will make a long version as I did with the Monte Carlo, I am more interested in guaranteeing functionality at the moment. I am going to run tests on the server after this push.